'Moved App to src/pages from components, defaultStyles tests'
