+++ package.json
@@ -64,6 +64,7 @@
 		"jest-styled-components": "^6.0.1",
 		"lint-staged": "^7.2.0",
 		"prettier": "^1.14.2",
+		"react-test-renderer": "^16.7.0",
 		"redux-devtools-extension": "^2.13.5",
 		"supertest": "^3.1.0",
 		"uglifyjs-webpack-plugin": "^1.2.7",
