+++ src/Root.js
@@ -3,7 +3,7 @@ import ReactDOM from 'react-dom';
 import { ThemeProvider } from 'styled-components';
 import theme from './theme';
 
-import App from './components/App.js';
+import App from './pages/App.js';
 
 const Root = () => (
 	<ThemeProvider theme={theme}>
