+++ src/pages/App.test.js
